backend/web/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0